requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
filelock==3.9.0
Flask==3.0.0
fsspec==2023.12.2
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.0
numpy==1.26.3
packaging==23.2
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
sympy==1.12
tokenizers==0.15.1
# torch==2.1.2+cpu -i https://download.pytorch.org/whl/cpu
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.4.0
urllib3==2.1.0
Werkzeug==3.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
filelock==3.9.0
Flask==3.0.0
fsspec==2023.12.2
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jiraone==0.8.2
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.0
numpy==1.26.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
shapely==2.0.3
six==1.16.0
sympy==1.12
tokenizers==0.15.1
torch==2.1.2+cpu
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.4.0
urllib3==2.1.0
Werkzeug==3.0.1
